Refactor: hooks, func 목적에 따른 파일 분리
